:sparkles: Feat : 알림 수정 api 공통 일정 controller로 이동
